WIP: Splitting Core::ControllerHelpers up into modules
