+++ core/app/controllers/spree/checkout_controller.rb
@@ -2,7 +2,7 @@ module Spree
   # Handles checkout logic.  This is somewhat contrary to standard REST convention since there is not actually a
   # Checkout object.  There's enough distinct logic specific to checkout which has nothing to do with updating an
   # order that this approach is waranted.
-  class CheckoutController < BaseController
+  class CheckoutController < Spree::FrontendController
     ssl_required
 
     before_filter :load_order
