+++ core/app/controllers/spree/content_controller.rb
@@ -1,5 +1,5 @@
 module Spree
-  class ContentController < BaseController
+  class ContentController < Spree::FrontendController
     # Don't serve local files or static assets
     before_filter { render_404 if params[:path] =~ /(\.|\\)/ }
