+++ core/app/controllers/spree/products_controller.rb
@@ -1,5 +1,5 @@
 module Spree
-  class ProductsController < BaseController
+  class ProductsController < Spree::FrontendController
     before_filter :load_product, :only => :show
     rescue_from ActiveRecord::RecordNotFound, :with => :render_404
     helper 'spree/taxons'
