+++ core/lib/spree/core/controller_helpers/respond_with.rb
