+++ promo/lib/spree/promo/engine.rb
@@ -8,7 +8,7 @@ module Spree
         Dir.glob(File.join(File.dirname(__FILE__), '../../../app/**/*_decorator*.rb')) do |c|
           Rails.configuration.cache_classes ? require(c) : load(c)
 
-          Spree::BaseController.class_eval do
+          Spree::FrontendController.class_eval do
             # Include list of visited paths in notification payload hash
             def default_notification_payload
               { :user => try_spree_current_user, :order => current_order, :visited_paths => session[:visited_paths] }
